Fixing tags deletion

+++ app/controllers/topics_controller.rb
@@ -5,7 +5,7 @@ class TopicsController < ApplicationController
   skip_before_action :verify_authenticity_token if Rails.env.test?
   before_action :authenticate_member
   before_action :is_joined, only: [:show]
-  before_action :is_owner, only: [:destroy, :edit, :update]
+  before_action :is_owner, only: [:edit, :update]
 
   def show
     @topic = Topic.find(params[:id])
@@ -77,11 +77,14 @@ class TopicsController < ApplicationController
     @topic = Topic.find(params[:id])
     @room = @topic.room
 
-    delete_dir(@topic)
-
-    @topic.destroy
-
-    redirect_to room_path @room
+    if @room.owner == current_member
+      delete_dir(@topic)
+      @topic.destroy
+      redirect_to room_path @room
+    else
+      flash[:notice] = "You do not have permission to do this action"
+      redirect_to room_path(@room)
+    end
   end
 
   def update_current_slide
